Move tests to another file
